refactor: delete InputConsoleView
